requirements.txt
alembic==1.0.5
blinker==1.4
Bootstrap-Flask==1.0.9
Click==7.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-DebugToolbar==0.10.1
Flask-Dropzone==1.5.3
Flask-Migrate==2.3.1
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
numpy==1.15.4
pandas==0.23.4
Pillow==5.4.1
PyMySQL==0.9.3
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.3
pytz==2018.9
six==1.12.0
SQLAlchemy==1.2.16
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
xlrd==1.2.0
